requirements.txt
flask
PyPDF2
openai
# langchain
# langchain-openai
python-dotenv